Extract git changes into separate class, expose root path option.
