+++ bake/covered/git.rb
@@ -7,15 +7,16 @@ def initialize(...)
 	
 	require 'set'
 	require 'rugged'
-	@repository = nil
 end
 
 # bake load_coverage_from_simplecov git:coverage
+# @parameter root [String] the root directory of the git repository.
 # @parameter branch [String] the branch to compare against.
 # @parameter input [Covered::Policy] the input policy to use.
-def statistics(branch: self.default_branch, input:)
+def statistics(root: context.root, branch: self.default_branch, input:)
 	input ||= context.lookup("covered:policy:current").call
-	modifications = lines_modified(branch)
+	changes = Covered::Git::BranchChanges.new(root)
+	modifications = changes.lines_modified(branch)
 	
 	# Calculate statistics:
 	statistics = Covered::Statistics.new
@@ -28,35 +29,4 @@ def statistics(branch: self.default_branch, input:)
 	end
 	
 	return statistics
-end
-
-private
-
-def repository(root = context.root)
-	@repository ||= Rugged::Repository.discover(root)
-end
-
-def default_branch
-	"main"
-end
-
-# Compute the lines modified for a given branch, returning a hash of paths to a set of line numbers.
-# @returns [Hash(String, Set(Integer))]
-def lines_modified(branch)
-	result = Hash.new{|k,v| k[v] = Set.new}
-	
-	diff = repository.diff(repository.rev_parse(branch), repository.last_commit)
-	
-	diff.each_patch do |patch|
-		path = patch.delta.new_file[:path]
-		
-		patch.each_hunk do |hunk|
-			hunk.each_line do |line|
-				result[path] << line.new_lineno if line.addition?
-			end
-		end
-	end
-
-	result.default = nil
-	return result.freeze
 end
